backend/requirements.txt
Flask>=2.0
psycopg2-binary>=2.9
python-dotenv>=0.21
flask-cors>=3.0
supabase
gunicorn
PyJWT>=2.8.0
requests>=2.28.0
reportlab>=4.0
Pillow>=9.0
matplotlib>=3.5
